requirements.txt
appdirs==1.4.4
arrow==0.15.8
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-adminlte3==0.1.6
django-appconf==1.0.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-googledrive-storage==1.5.2
django-heroku==0.3.1
django-inline-actions==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
oauth2client==4.1.3
pep8==1.7.1
Pillow==7.2.0
pipenv==2020.6.2
portalocker==2.0.0
protobuf==3.12.4
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.29
virtualenv-clone==0.5.4
whitenoise==5.1.0
wrapt==1.12.1
